User History Facade
This is the work I am going to do to allow for entry of user history
about eating habits. The UserHistory object in Facade is what will be
delivered to the Display as information about the user history. I also
added a table to story history items in the database.