Added code to check caloryGoals
Changed the userProfiles table to include a caloryGoal column
userHistory is the users history of food. Will eventually link that to
the userProfiles table.
userGoalAccomplish is just a bit of code that checks the caloryGoal and
compares it to how many calories the user has consumed.

The new tables have some test rows for you to try out.